+++ src/global_align.jl
@@ -7,7 +7,7 @@
 
 """
     global_align(seqA::String, seqB::String,
-        sub_header::Dict{Char,Int}, sub_matrix::Array{Int,2},
+        sub_header::Dict{Char,Int}, sub_matrix::Array{Float64,2},
         gap_open::Int=-12, gap_extend::Int=-4,
         end_gap_open::Int=0, end_gap_extend::Int=0)
   
@@ -22,7 +22,7 @@ very short overlap alignments if the sequences are very dissimilar.
 Only returns one best-scoring alignment, regardless of ties for best score.
 """
 function global_align(seqA::String, seqB::String,
-                      sub_header::Dict{Char,Int}, sub_matrix::Array{Int,2};
+                      sub_header::Dict{Char,Int}, sub_matrix::Array{Float64,2};
                       gap_open::Int=-12, gap_extend::Int=-4,
                       end_gap_open::Int=0, end_gap_extend::Int=0)
